requirements.txt
flask
requests
pandas
plotly
python-dotenv
scikit-learn
numpy
flask-wtf
email_validator
sqlalchemy
alembic
tenacity
tzdata
xgboost
flask-login
pynostr
pymongo